requirements.txt
aiogrpcclient==1.3.0
aiohttp==3.8.3
aiokit==1.2.0
aiosignal==1.3.1
aiosumma==2.28.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
fire==0.5.0
frozenlist==1.3.3
grpcio==1.51.1
idna==3.4
izihawa-utils==1.1.0
multidict==6.0.4
orjson==3.8.5
protobuf==4.21.12
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
termcolor==2.2.0
urllib3==1.26.14
uvloop==0.17.0
yarl==1.8.2
